vector_similarity_fcts: add calculating similarity between vectors
This commit introduces a `compute_similarity` function overriden
by each similarity function using their expected metric to calculate
the similarity between given vectors.

The implementation of those calculations is inspired by:
https://github.com/unum-cloud/USearch/blob/a2f17599101729d667dc0260dd278852d9098183/include/usearch/index_plugins.hpp#L1304-L1385
but changed to return the similarity instead of the distance to match
Cassandra's implementation.

The similarity metrics shows how similar the vectors are
in range from 0 - not similar at all to 1 - the same.